requirements.txt
streamlit
google-generativeai
feedparser
beautifulsoup4
python-dateutil
